refactor: rename tasks & tests files
