+++ src/common-character-count.js
