+++ src/delete-digit.js
