+++ src/encode-line.js
