+++ src/matrix-elements-sum.js
