+++ src/mine-sweeper.js
