+++ test/encode-line.test.js
@@ -1,12 +1,12 @@
 const { assert } = require('chai');
 const { testOptional } = require('../extensions/index.js');
-const { encodeLine } = require('../src/st-encode-line.js');
+const { encodeLine } = require('../src/encode-line.js');
 
 it.optional = testOptional;
 
 Object.freeze(assert);
 
-describe('st-encode-line', () => {
+describe('Encode line', () => {
   it.optional('should return encoding version of string', () => {
     assert.strictEqual(encodeLine('aaaatttt'), '4a4t');
     assert.strictEqual(encodeLine('aabbccc'), '2a2b3c');
