+++ test/mine-sweeper.test.js
@@ -1,12 +1,12 @@
 const { assert } = require('chai');
 const { testOptional } = require('../extensions/index.js');
-const { minesweeper } = require('../src/st-mine-sweeper.js');
+const { minesweeper } = require('../src/mine-sweeper.js');
 
 it.optional = testOptional;
 
 Object.freeze(assert);
 
-describe('st-mine-sweeper', () => {
+describe('Minesweeper', () => {
   it.optional('should return minesweeper game setup', () => {
     assert.deepEqual(
       minesweeper([
