+++ test/sum-digits.test.js
@@ -1,12 +1,12 @@
 const { assert } = require('chai');
 const { testOptional } = require('../extensions/index.js');
-const { getSumOfDigits } = require('../src/st-sum-digits.js');
+const { getSumOfDigits } = require('../src/sum-digits.js');
 
 it.optional = testOptional;
 
 Object.freeze(assert);
 
-describe('st-sum-digits', () => {
+describe('Sum digits', () => {
   it.optional('should return the sum of digits', () => {
     assert.strictEqual(getSumOfDigits(91), 1);
     assert.strictEqual(getSumOfDigits(100), 1);
